+++ core/persist/index.ts
@@ -1,9 +1,9 @@
 export * from './Persistable';
 export * from './PersistOptions';
 export * from './PersistenceProvider';
-export * from './LocalStorageProvider';
-export * from './SessionStorageProvider';
-export * from './DashViewProvider';
-export * from './PrefProvider';
-export * from './CustomProvider';
-export * from './viewmanager/ViewManagerProvider';
+export * from './provider/LocalStorageProvider';
+export * from './provider/SessionStorageProvider';
+export * from './provider/DashViewProvider';
+export * from './provider/PrefProvider';
+export * from './provider/CustomProvider';
+export * from './provider/ViewManagerProvider';
