+++ core/persist/provider/DashViewProvider.ts
@@ -7,7 +7,7 @@
 
 import type {DashViewModel} from '@xh/hoist/desktop/cmp/dash'; // Import type only
 import {throwIf} from '@xh/hoist/utils/js';
-import {PersistenceProvider, PersistenceProviderConfig} from './';
+import {PersistenceProvider, PersistenceProviderConfig} from '../PersistenceProvider';
 
 /**
  * PersistenceProvider that stores state within a DashViewModel.
